// src/type/value.type.ts
/**
 * @author Chris Humboldt
 */

export interface ClassValueOptions {
  applyClass: any;
  applyClassToMap: boolean;
  data: any;
}
export interface ValueOptions {
  applyClass?: any;
  applyClassToMap?: boolean;
  fallback?: any;
  transform?: any;
  data: any;
}
export interface SetValueOptions {
  fallback?: any;
  data: any;
}